Add a 'Developer Mode' to the RPC Server
In an attempt to promote 'eating our own dog food', this change would
allow us to run the RPC Server - and in turn, CPRS - against
(non-clinical) prototyped RPC Locker and VDM models while being as
non-invasive as possible to the overall production demo functionality.

To activate "Developer Mode", a user would need to tweak the RPC Server
configuration file by adding a 'developerMode' flag set to 'true'. This
causes the queue worker to load the RPC locker and VDM models from the
NonClinicalRPCs repo, assuming a VDP developer directory structure.

Without the explicitly-set 'developerMode' flag in the configuration,
it will be as if the code block was never written.